+++ templates/base.html
@@ -186,38 +186,14 @@
                     </div>
                 </li>
                 {% if user.is_authenticated %}                    
-                <li class="bg-secondary list-group-item list-group-item-action" style="background-color: #191919!important;">
+                <a class="bg-secondary list-group-item list-group-item-action"
+                   href="{% url 'home' %}"
+                   style="background-color: #191919!important;color:#b1b1b1!important;">
                     <div class="d-flex w-100 justify-content-start align-items-center">
-                        <a class="d-flex w-100 justify-content-start align-items-center dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false"
-                            style="color:#b1b1b1!important;">
-                            <span class="fa-solid fa-bars fa-xl" style="margin-right:5px;"></span>
-                            <span class="menu-collapsed" style="margin-right:5px;">MENU</span>
-                        </a>
-                        <ul class="bg-secondary dropdown-menu"  aria-labelledby="navbarDropdown">
-                            <li>
-                                <a href="{% url 'examSelect' %}" class="bg-secondary list-group-item list-group-item-action">
-                                        <i class="fas fa-folder-open" style="margin-right:5px;"></i>Open exam</a>
-                            </li>                           
-                            {% if user|more_arg:None|is_allowed:'creation' %}
-                             <div class="bg-secondary dropdown-divider"></div>
-                            <li>
-                                <a class="bg-secondary list-group-item list-group-item-action" href="{% url 'create_exam_project' %}">
-                                    <i class="fas fa-plus" style="margin-right:5px;"></i>Create exam</a>
-                            </li>
-                            {% endif %}
-                            <div class="bg-secondary dropdown-divider"></div>
-                            <li>
-                                <a href="{% url 'generate_room_plan' %}" class="bg-secondary list-group-item list-group-item-action">
-                                    <i class="fa-solid fa-meteor" style="margin-right:5px;color:#fdfdfd;"></i>Room plan</a>       
-                            </li>
-                            <div class="bg-secondary dropdown-divider"></div>
-                            <li>
-                                <a href="{% url 'csvgen' %}" class="bg-secondary list-group-item list-group-item-action">
-                                    <i class="fa-solid fa-file-excel" style="margin-right:5px;color:#fdfdfd;"></i>Students CSV generator</a>       
-                            </li>
-                      </ul>
+                        <span class="fa-solid fa-table-columns fa-xl" style="margin-right:5px;"></span>
+                        <span class="menu-collapsed" style="margin-right:5px;">DASHBOARD</span>
                     </div>
-                </li>   
+                </a>
                 {% if exam_selected %}
                     <a {% if user|more_arg:exam_selected|is_allowed:'preparation' %} href="{% url 'examInfo' exam_selected.pk %}" {% endif %}class="bg-secondary list-group-item list-group-item-action sidebar-separator-title font-weight-bold" style="height:auto;background-color: #1d2124;margin-top:20px;">
                         <span title="{{ exam_selected.code }}" class="fa-solid fa-book-open" style="margin-right: 5px;"></span>
@@ -322,7 +298,6 @@
                                 <span class="menu-collapsed" style="color:#CCA71CFF;margin-right:5px;">Info</span>
                             </a>
                             <ul class="bg-secondary dropdown-menu"  aria-labelledby="navbarDropdown">
-                                {%  if exam_selected.scales.all %}
                                 <li><a class="bg-secondary dropdown-item" href="{% static 'docs/html/index.html' %}" style="padding-top:15px;"><i class="fa-solid fa-circle-question fa-lg" style="margin-right:5px;"></i>Documentation</a></li>
                                 <li><a class="bg-secondary dropdown-item"
                                        type="button"
@@ -330,7 +305,6 @@
                                           data-bs-target="#aboutModal"
                                        style="padding-top:15px;">
                                     <i class="fa-solid fa-circle-info fa-lg" style="margin-right:5px;"></i>About</a></li>
-                                {% endif %}
                           </ul>
                         </div>
                     </li>
